+++ Examples/CQRS/Examples.Cqrs/Commands/SimplePostCommand.cs
@@ -1,21 +1,23 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 using APIBlox.AspNetCore.Contracts;
+using APIBlox.AspNetCore.Types;
 using APIBlox.NetCore.Attributes;
 using Examples.Resources;
 
 namespace Examples.Commands
 {
     [InjectableService]
-    internal class SimplePostCommand : ICommandHandler<ExampleRequestObject>
+    internal class SimplePostCommand : ICommandHandler<ExampleRequestObject,HandlerResponse>
     {
-        public Task HandleAsync(ExampleRequestObject requestCommand, CancellationToken cancellationToken)
+        public Task<HandlerResponse> HandleAsync(ExampleRequestObject requestCommand, CancellationToken cancellationToken)
         {
             // This implementation does NOT return anything (other than the task object obviously).
 
-            var foo = requestCommand.SomeValue;
+            var res = new HandlerResponse().SetErrorToBadRequest("u blow");
+            res.Result = requestCommand.SomeValue;
 
-            return Task.CompletedTask;
+            return Task.FromResult(res);
         }
     }
 }