+++ Examples/CQRS/Examples.Cqrs/Configuration/DecoratorsConfig.cs
@@ -14,12 +14,12 @@ namespace Microsoft.Extensions.DependencyInjection
                 //
                 // Nothing more than a stopwatch.  Ideally (my opinion) an API
                 // call should never take more than 3 seconds.
-                typeof(MetricsCommandHandlerDecorator<>),
+                typeof(MetricsCommandHandlerDecorator<,>),
 
                 //
                 // This one is still a work in progress and is most likely
                 // more valuable when combined with domain events.
-                typeof(TransactionScopeCommandHandlerDecorator<>),
+                typeof(TransactionScopeCommandHandlerDecorator<,>),
                 typeof(SimplePostCommandDecorator)
             );